Added Solid Rects
I've implemented a system that allows for rects to repel the player when
they come in contact, simulating a solid structure. It is currently very
buggy, and only implemented for the tower. I'll be cleaning this up a
LOT in the future, and simply adding anything that should repel the
player to a list of rects. It is very possible to get stuck in the tower
at the moment.